implement router for page query
