+++ frontend/package.json
@@ -13,6 +13,7 @@
     "classnames": "^2.3.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.0",
     "react-test-renderer": "^17.0.2",
     "styled-components": "^5.3.5",
